pyproject.toml
[tool.poetry]
name = "parser_whales_tracker"
version = "0.1.0"
description = ""
authors = ["DMITRII <dengivseti@list.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dotenv = "^0.17.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
